ACC-2104 let curies link relation render as array when there are embedded resources
spring-hateoas automatically adds the curies link relation on resources with curied links and also makes sure this curies link relation is rendered as an array. But spring-hateoas doesn't add the curies link relation when there are embedded resources with curied links, we add them manually with SpringDataEmbeddedCuriesResourceProcessor, but this wasn't rendering the link relation as an array.