[ACC-2149] Inline and remove Entity#getFilterByPath() method
This method was unsuitable to put on the model, because it returns the first AttributeSearchFilter, without taking into account the specific type.
There can be more than one SearchFilter on the same AttributePath. In the case in the XToManyRelationRestController, we specifically
want the ExactSearchFilter and nothing else, so we should just do the filtering in code there.