[ACC-2050] Introduce domain values for EntityId (and EntityIdentity/Version)
Next to purely the entity ID, we have a broader class EntityIdentity, which is a full specification of the identity of an entity at a point in time
The EntityIdentity contains the EntityName, EntityId and Version.

Version in its turn is potentially a specification of the exact entity version that the identity refers to.
It is also possible for a Version to be unspecified, which means that either no specific version is requested,
or that the storage does not have a specific version available