ACC-2200 fix previous link and prev_cursor when on second page
For the first page, the cursor is null and the url does not contain '_cursor' query parameter. The second page contains a link to the first page called 'prev' and a reference to the previous cursor called 'prev_cursor' in the page metadata (or absent if there is no previous page). The 'prev' link gets constructed using the 'self' link, but '_cursor' was not removed from the link. 'prev_cursor' was absent for both first and second page, now it is null when the previous page has no cursor and absent when there is no previous page.