[ACC-2210] Replace ContentWriter that has an OutputStream with ContentStore#writeContent() receiving an InputStream
For the S3 implementation, it turns out that having an InputStream to
work with is much easier than writing to an OutputStream (from an
InputStream outside the ContentStore API), and then having to convert it
back to an InputStream to write it to S3. That takes writing to memory
buffers multiple times.

Additionally, a lot of implementation becomes easier when you don't have
to take action based on the OutputStream being closed (e.g. for
encryption), and can just put that code directly after the call to the
wrapped ContentStore#writeContent.