[ACC-2269] Use actual implementations for RelationRestControllerTest
Replace the unit test for RelationRestController with a functional test

Unit-testing the controllers is not very useful, because it is a simple pass-through to the domain APIs.
Instead, the interesting things to test are whether the interactions are correct, and that the REST API performs as expected under all scenarios.