[ACC-2269] Handle transactions for QueryEngine manually
There are dangerous gotchas by using the `@Transactional` annotation on
the implementation class only.

Default methods on the `QueryEngine` interface will not run inside a
transaction at all.
This happens because:
1. the interceptor does not intercept those methods directly (as they
   are not annotated)
2. calls to `this` are never intercepted by interceptors (because they
   are dispatched internally, and don't pass the proxy object that
intercepts methods)

The result is that adding default methods to `QueryEngine` is dangerous,
because it bypasses transaction management.
We explicitly rely on rolling back transactions for permission
management. Being non-transactional would mean that changes will go
through, even though a permission exception is thrown to revert the
change.