Ensure that xor can never be no encryption
The encryption tests check that content is encrypted. XOR'ing with 0 will result in no encryption being applied at all, resulting in the tests failing.
The chance of having a 0 byte is 1/256, which means that same chance
that the test will fail. Having a flakey test like that should be
avoided, so ensure that a new value is selected when the byte is 0.