[ACC-2378] Refactor BlindRelationOverwriteException to not use target endpoint name
A BlindRelationOverwriteException is conceptually about the relation on
the other side already pointing to an existing value, but the inverse of
a relation doesn't always exist like that (for uni-directional
relations), and we wouldn't be able to represent something consistent
and useful.

Instead, flip the exception around, where an existing value and the new
relation can't both reference a target entity.