[ACC-2439] Automatically generate search filters for to-many relations
Instead of failing the creation of an application model when required
search filters are missing for to-many relations, generate them
directly.

This makes it easier to build application models in tests, without
requiring additional code there to ensure that the necessary search
filters are present.
It also makes it easier to create json application models, as you don't
need to think about the internal filters that are required for the
application to operate correctly.