[ACC-2439] Fix execution of interference
Usage of setExecutor() in the interference thread results in the thread trying to self-join, which hangs forever.
Due to this, no interference code was actually being executed in the tests. The 1-second timeout thread join would
log a warning and only the main thread would continue executing.
Because the executor was already replaced with a NOOP executor, there was no error that the interference thread was not running