[ACC-2439] Fix interference running early breaking query collection (due to exception thrown early)
In a lot of cases, the interference code that runs will change the
database state. Due to this, most of the test code will throw an
exception early (e.g. version mismatch or entity does not exist anymore)
This causes only the first select query to be collected, and the
concurrency tests not testing most of the interesting interference
scenarios.

To fix this, during the pre-run, run the interference code _after_ the
test code has run instead of before. Because it runs later, it can't
affect the execution of the test code in any way, but the interference
code is still run before verification is performed. That way,
verification during the pre-run can also assert on state dependent on
interference code having been run at some point in time.